+++ cmodules/msgPayloadDefC/GpsOutMsgPayload.h
@@ -8,9 +8,9 @@
 /*! @brief Brief description of what this message contains */
 typedef struct {
     int state;
-    float lat;
-    float lon;                  // 0 if no plume detected | 1 if plume detected
-    float altitude;
+    double lat;
+    double lon;                  // 0 if no plume detected | 1 if plume detected
+    double altitude;
 }GpsOutMsgPayload;
 
 #endif
